requirements.txt
pyopenssl
gitpython
docker
behave
pyhamcrest
pyaml
